Overhauls the data model to use ManyToMany relationships between Disease and Vaccine entities, removing the intermediate DiseaseVaccine entity.
Deletes obsolete controller, service, mapper, and repository files related to the previous explicit many-to-many implementation.
Adjusts entity relationships to streamline data access and simplifies management of disease-vaccine associations.
Removes unnecessary service autowiring in PupilController.

+++ src/main/java/com/schoolhealth/schoolmedical/entity/DiseaseVaccineId.java
@@ -12,7 +12,7 @@ import java.util.Objects;
 @EqualsAndHashCode
 public class DiseaseVaccineId implements Serializable {
 
-    private Long diseaseId;
-    private Long vaccineId;
+    private int diseaseId;
+    private int vaccineId;
 
 }
